Do not allow install doom files on a target that doesn't have a doom.rock plugin.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@14697 a1c6a512-1295-4272-9138-f99709370657

+++ rbutil/rbutilqt/rbutilqt.cpp
@@ -677,13 +677,19 @@ void RbUtilQt::installVoice()
 
 void RbUtilQt::installDoomBtn()
 {
+    QFile doomrock(userSettings->value("mountpoint").toString()+"/.rockbox/rocks/games/doom.rock");
+    if(!doomrock.exists()){
+        QMessageBox::critical(this, tr("Error"), tr("Your device doesn't have a doom plugin. Aborting."));
+        return;
+    }
+
     if(QMessageBox::question(this, tr("Confirm Installation"),
            tr("Do you really want to install the game addon files?"),
            QMessageBox::Yes | QMessageBox::No) != QMessageBox::Yes) return;
     // create logger
     logger = new ProgressLoggerGui(this);
     logger->show();
-    
+
     installDoom();
 }
 bool RbUtilQt::installDoomAuto()
